examples/traffic_lights/requirements.txt
tensorflow-gpu
opencv-python
eclipse-zenoh